different work stealing interface for prep and DrvX
