+++ pando-rt/src/prep/cores.hpp
@@ -7,10 +7,12 @@
 #include <cstddef>
 #include <tuple>
 
+#include "../start.hpp"
 #include "../queue.hpp"
 #include "pando-rt/execution/task.hpp"
 #include "pando-rt/index.hpp"
 #include "pando-rt/status.hpp"
+#include <pando-rt/benchmark/counters.hpp>
 
 namespace pando {
 
@@ -120,6 +122,11 @@ public:
    * @brief Returns a flag to check if the core is active.
    */
   static CoreActiveFlag getCoreActiveFlag() noexcept;
+
+  /**
+   * @brief Steals work from other cores.
+   */
+  static void workStealing(std::optional<pando::Task>& task, SchedulerFailState& failState, counter::Record<std::int64_t>& idleCount, counter::HighResolutionCount<IDLE_TIMER_ENABLE>& idleTimer);
 };
 
 } // namespace pando
