+++ pando-rt/src/start.hpp
@@ -4,6 +4,17 @@
 #ifndef PANDO_RT_SRC_START_HPP_
 #define PANDO_RT_SRC_START_HPP_
 
+#include <cstdint>
+
+constexpr bool IDLE_TIMER_ENABLE = false;
+
+constexpr std::uint64_t STEAL_THRESH_HOLD_SIZE = 16;
+
+enum SchedulerFailState{
+  YIELD,
+  STEAL,
+};
+
 /**
  * @brief Start function for each hart.
  *
